gateway/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP client for proxying
httpx>=0.26.0
websockets>=12.0

# Validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# JWT validation
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
